services/ai-service/requirements.txt
fastapi
uvicorn[standard]
prophet
redis
python-dotenv
pandas
pydantic
